Added input reader
The InputActions asset has been turned into a script by Unity and is now
used to let the InputReader inherit the input actions from the
InputSystem.
The InputReader is able to be referenced as an instance without needing
to be used in a scene because i made it a scriptable object.