AllEars movement and aggro
-Allears can move
It will move to random locations around the room

-AllEars can get aggro'd
When the aggro function gets called (which doesnt happen rn) AllEars moves towards the location given at a fast speed, and attacks there.
If a new sound is made during the walking it will redirect
If a new sound is made during attacking it will finish its attack and move to the new spot and attack again.
These sounds do not stack, it will go to the newest sound.

-Monsters all have aggro and deaggro functions

-Changed the color of the ear material