Added player interaction
Added an interactor that can shoots a raycast from the player which can call the Hover() function of the collided object when its an interactable object showing an outline.
When hovering over an interactable object and pressing E it calls the interact object's Interact function which calls a customizable Unity Event.