Added noise function, added player watch check
-Added a static function to make noise
This makes all the all ears go to where the noise came from

-Added a check for when the player is watching
If a player has an object in a second camera on them, it will return true. Current issue: it looks through walls.

-Added the stalkers base attacking
It deals insanity when you look at it

-Added a temp folder
for the inverted sphere model and anything else that can later be removed.