Can now select slots and call the use function of an item
You can now use 0 - 9 on keyboard to navigate the hotbar. The hotbar only shows slots with items and removes the item from the hotbar when the use function is called of that item.
All items are stored as scriptable objects and their use function can be overidden in InventoryItem.cs.

You currently cannot pickup or drop items.